% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{add_choices}
\alias{add_choices}
\title{Title}
\usage{
add_choices(
  choices_meta,
  backup_dir = NULL,
  overwrite = FALSE,
  append = FALSE,
  open = FALSE
)
}
\arguments{
\item{choices_meta}{Data}

\item{backup_dir}{Filepath}

\item{overwrite}{Boolean}

\item{append}{Boolean}

\item{open}{Boolean}
}
\value{
Used for side effects
}
\description{
Title
}
\examples{
\dontrun{

}
}
