% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{add_data_context_output}
\alias{add_data_context_output}
\title{Title}
\usage{
add_data_context_output(app_type, data_prefix, month_col)
}
\arguments{
\item{app_type}{Character}

\item{data_prefix}{Character}

\item{month_col}{Character}
}
\value{
Used for side effects
}
\description{
Title
}
\examples{
\dontrun{

}
}
