% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{add_levels}
\alias{add_levels}
\title{Title}
\usage{
add_levels(
  app_type,
  data,
  levels_meta,
  backup_dir = NULL,
  overwrite = FALSE,
  open = FALSE
)
}
\arguments{
\item{app_type}{Character}

\item{data}{Data}

\item{levels_meta}{Data}

\item{backup_dir}{Filepath}

\item{overwrite}{Boolean}

\item{open}{Boolean}
}
\value{
Used for side effects
}
\description{
Title
}
\examples{
\dontrun{

}
}
