% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{add_output}
\alias{add_output}
\title{Title}
\usage{
add_output(app_type, page, output_name, render_func, output_func, args)
}
\arguments{
\item{app_type}{Character}

\item{page}{Filepath}

\item{output_name}{Character}

\item{render_func}{Function}

\item{output_func}{Function}

\item{args}{Vector of characters}
}
\value{
Used for side effects
}
\description{
Title
}
\examples{
\dontrun{

}
}
