% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{add_page}
\alias{add_page}
\title{Title}
\usage{
add_page(
  app_type,
  page_name,
  page_menu_name = page_name,
  backup_dirs = NULL,
  overwrite = FALSE,
  open = FALSE
)
}
\arguments{
\item{app_type}{Character}

\item{page_name}{Character}

\item{page_menu_name}{Character}

\item{backup_dirs}{Vector of filepaths}

\item{overwrite}{Boolean}

\item{open}{Boolean}
}
\value{
Used for side effects
}
\description{
Title
}
\examples{
\dontrun{

}
}
