% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{add_reactive_data}
\alias{add_reactive_data}
\title{Title}
\usage{
add_reactive_data(app_type, data_prefix, month_col, region_col)
}
\arguments{
\item{app_type}{Character}

\item{data_prefix}{Character}

\item{month_col}{Character}

\item{region_col}{Vector of filepaths}
}
\value{
Used for side effects
}
\description{
Title
}
\examples{
\dontrun{

}
}
