% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{add_survey_doc_output}
\alias{add_survey_doc_output}
\title{Title}
\usage{
add_survey_doc_output(app_type, data_prefix, survey_doc)
}
\arguments{
\item{app_type}{Character}

\item{data_prefix}{Character}

\item{survey_doc}{Character}
}
\value{
Used for side effects
}
\description{
Title
}
\examples{
\dontrun{

}
}
