% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{add_ui}
\alias{add_ui}
\title{Title}
\usage{
add_ui(
  page,
  section,
  label,
  output_name,
  ui_func,
  section_num,
  use_comment_group,
  use_aspect_radio
)
}
\arguments{
\item{page}{Character}

\item{section}{Character}

\item{label}{Character}

\item{output_name}{Character}

\item{ui_func}{Function}

\item{section_num}{Int}

\item{use_comment_group}{Boolean}

\item{use_aspect_radio}{Boolean}
}
\value{
Used for side effects
}
\description{
Title
}
\examples{
\dontrun{

}
}
