% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aspect_radio_select}
\alias{aspect_radio_select}
\title{Create Shiny input for selecting from multiple similar aspects in stacked
vertical charts}
\usage{
aspect_radio_select(id, choices)
}
\arguments{
\item{id}{Character}

\item{choices}{Vector of characters}
}
\value{
HTML
}
\description{
Create Shiny input for selecting from multiple similar aspects in stacked
vertical charts
}
\examples{
aspect_radio_select("help-used", c("Helpdesk", "Online"))
}
