% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{check_overwrite}
\alias{check_overwrite}
\title{Title}
\usage{
check_overwrite(files, backup_dirs, overwrite)
}
\arguments{
\item{files}{Vector of filepaths}

\item{backup_dirs}{Vector of filepaths}

\item{overwrite}{Boolean}
}
\value{
Used for side effects
}
\description{
Title
}
\examples{
\dontrun{}
}
