% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{coding_horizontal_bar}
\alias{coding_horizontal_bar}
\title{Use for displaying \% of coded responses.}
\usage{
coding_horizontal_bar(
  data,
  coded_column,
  total_column,
  colour = "#005EB8",
  chart_title = NULL,
  arrange_desc = TRUE,
  low_base = 100,
  top_ranks = 7
)
}
\arguments{
\item{data}{Data}

\item{coded_column}{Character}

\item{total_column}{Character}

\item{colour}{Character}

\item{chart_title}{Character}

\item{arrange_desc}{Boolean}

\item{low_base}{Int}

\item{top_ranks}{Int}
}
\value{
Highcharts object
}
\description{
Use for displaying \% of coded responses.
}
\examples{
\dontrun{}
}
