% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{colour_ramp}
\alias{colour_ramp}
\title{Make colour ramp palette. Default start colour is NHS Dark Blue. To allow for
avoiding having white as last colour, n_colours + 1 are created.}
\usage{
colour_ramp(
  n_colours,
  first = "#003087",
  last = "#FFFFFF",
  include_white = FALSE
)
}
\arguments{
\item{n_colours}{Int}

\item{first}{Character}

\item{last}{Character}

\item{include_white}{Boolean}
}
\value{
colorRampPalette
}
\description{
Make colour ramp palette. Default start colour is NHS Dark Blue. To allow for
avoiding having white as last colour, n_colours + 1 are created.
}
\examples{
\dontrun{

}
}
