% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{content_box}
\alias{content_box}
\title{Wrap given tags in a content box}
\usage{
content_box(title, ...)
}
\arguments{
\item{title}{Character}

\item{...}{List of HTML tags}
}
\value{
HTML
}
\description{
Wrap given tags in a content box
}
\examples{
\dontrun{}
}
