% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{create_app_r}
\alias{create_app_r}
\title{Title}
\usage{
create_app_r(
  app_type,
  app_title,
  data_prefix,
  month_col,
  region_col,
  backup_dir = NULL,
  overwrite = FALSE,
  open = FALSE
)
}
\arguments{
\item{app_type}{Character}

\item{app_title}{Character}

\item{data_prefix}{Character}

\item{month_col}{Character}

\item{region_col}{Character}

\item{backup_dir}{Filepath}

\item{overwrite}{Boolean}

\item{open}{Boolean}
}
\value{
Used for side effects
}
\description{
Title
}
\examples{
\dontrun{

}
}
