% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{create_choices}
\alias{create_choices}
\title{Title}
\usage{
create_choices(choices_meta)
}
\arguments{
\item{choices_meta}{Data}
}
\value{
Vector of characters
}
\description{
Title
}
\examples{
\dontrun{}
}
