% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{create_levels}
\alias{create_levels}
\title{Title}
\usage{
create_levels(data, levels_meta)
}
\arguments{
\item{data}{Data}

\item{levels_meta}{Data}
}
\value{
Vector of characters
}
\description{
Title
}
\examples{
\dontrun{}
}
