% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{create_output}
\alias{create_output}
\title{Title}
\usage{
create_output(app_type, output_name, render_func, output_func, args)
}
\arguments{
\item{app_type}{Character}

\item{output_name}{Character}

\item{render_func}{Function}

\item{output_func}{Function}

\item{args}{Character}
}
\value{
Used for side effects
}
\description{
Title
}
\examples{
\dontrun{

}
}
