% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{create_ui_outputs_r}
\alias{create_ui_outputs_r}
\title{Title}
\usage{
create_ui_outputs_r(
  app_type,
  data_prefix,
  page_1,
  survey_doc,
  backup_dir = NULL,
  overwrite = FALSE,
  open = FALSE
)
}
\arguments{
\item{app_type}{Character}

\item{data_prefix}{Character}

\item{page_1}{Character}

\item{survey_doc}{Character}

\item{backup_dir}{Filepath}

\item{overwrite}{Boolean}

\item{open}{Boolean}
}
\value{
Used for side effects
}
\description{
Title
}
\examples{
\dontrun{

}
}
