% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{date_Ymd_to_bY}
\alias{date_Ymd_to_bY}
\title{Set format of date to abbreviated month and full year.}
\usage{
date_Ymd_to_bY(date_to_format)
}
\arguments{
\item{date_to_format}{Date}
}
\value{
Character
}
\description{
Set format of date to abbreviated month and full year.
}
\examples{
date_Ymd_to_bY(as.Date("2023/03/23"))
}
