% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{decimal_places}
\alias{decimal_places}
\title{Set number of decimal places of number and return as char.}
\usage{
decimal_places(x, k)
}
\arguments{
\item{x}{Int}

\item{k}{Int}
}
\value{
Character
}
\description{
Set number of decimal places of number and return as char.
}
\examples{
\dontrun{

}
}
