% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{file_insert_lines}
\alias{file_insert_lines}
\title{Title}
\usage{
file_insert_lines(.file, .lines, .match = NA, .nth_match = 1, open = FALSE)
}
\arguments{
\item{.file}{Filepath}

\item{.lines}{Vector of characters}

\item{.match}{Character or NA}

\item{.nth_match}{Int}

\item{open}{Boolean}
}
\value{
Used for side effects
}
\description{
Title
}
\examples{
\dontrun{

}
}
