% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{file_remove_line}
\alias{file_remove_line}
\title{Title}
\usage{
file_remove_line(.file, .match = NA, open = FALSE)
}
\arguments{
\item{.file}{Filepath}

\item{.match}{Character or NA}

\item{open}{Boolean}
}
\value{
Used for side effects
}
\description{
Title
}
\examples{
\dontrun{}
}
