% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%AND\%}
\alias{\%AND\%}
\title{Helper that will return NULL if at least one of LHS and RHS is NA or NULL, or
it will return the RHS otherwise.}
\usage{
LHS \%AND\% RHS
}
\arguments{
\item{LHS}{Anything}

\item{RHS}{Anything}
}
\value{
RHS or NULL
}
\description{
Helper that will return NULL if at least one of LHS and RHS is NA or NULL, or
it will return the RHS otherwise.
}
\examples{
1 \%AND\% 1:3   # 1 2 3
NA \%AND\% 1:3 # NULL
}
