% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{group_table}
\alias{group_table}
\title{Use for free text tables such as comments and reasons; give the column of
interest and the associated selectInput id to use for NPS grouping as strings.}
\usage{
group_table(
  data,
  comment_column,
  month_col,
  group_select = NULL,
  comment_name_in_DT = "Reasons",
  data_prefix = NULL
)
}
\arguments{
\item{data}{Data}

\item{comment_column}{Character}

\item{month_col}{Character}

\item{group_select}{Character or NULL}

\item{comment_name_in_DT}{Character}

\item{data_prefix}{Character}
}
\value{
Datatable
}
\description{
Use for free text tables such as comments and reasons; give the column of
interest and the associated selectInput id to use for NPS grouping as strings.
}
\examples{
\dontrun{

}
}
