% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{horizontal_bar}
\alias{horizontal_bar}
\title{Use for horizontal barcharts showing percentages of some categorical
feature. Use for non-rating style questions with more than 3 responses only.}
\usage{
horizontal_bar(
  data,
  columns,
  cols_start_with = FALSE,
  responses,
  colour = "#005EB8",
  chart_title = NULL,
  arrange_desc = TRUE,
  low_base = 100,
  month_col
)
}
\arguments{
\item{data}{Data}

\item{columns}{Vector of characters}

\item{cols_start_with}{Boolean}

\item{responses}{Variable}

\item{colour}{Character}

\item{chart_title}{Character}

\item{arrange_desc}{Boolean}

\item{low_base}{Int}

\item{month_col}{Character}
}
\value{
Highcharts object
}
\description{
Use for horizontal barcharts showing percentages of some categorical
feature. Use for non-rating style questions with more than 3 responses only.
}
\examples{
\dontrun{

}
}
