% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{info_tooltip}
\alias{info_tooltip}
\title{Create info tooltip to be used with tippy.js}
\usage{
info_tooltip(
  content,
  style = "display: inline; padding-left: 1px; font-size: 1 rem; font-weight: 400"
)
}
\arguments{
\item{content}{HTML elements}

\item{style}{Character}
}
\value{
HTML
}
\description{
Create info tooltip to be used with tippy.js
}
\examples{
\dontrun{}
}
