% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{insert_lines_at}
\alias{insert_lines_at}
\title{Title}
\usage{
insert_lines_at(lines, to_insert, insert_at)
}
\arguments{
\item{lines}{Vector of characters}

\item{to_insert}{Vector of characters}

\item{insert_at}{Int}
}
\value{
Vector of characters
}
\description{
Title
}
\examples{
\dontrun{}
}
