% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{label}
\alias{label}
\title{Style text as label}
\usage{
label(label_text, ...)
}
\arguments{
\item{label_text}{Character}

\item{...}{HTML elements}
}
\value{
HTML
}
\description{
Style text as label
}
\examples{
\dontrun{

}
}
