% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{line_chart}
\alias{line_chart}
\title{Use for NPS and NES line charts over time (monthly and 3 month rolling
average); measure should be either "NPS" or "NES"}
\usage{
line_chart(
  data,
  measure_column,
  month_col,
  three_month_rolling = FALSE,
  last_valid_month_range_selection,
  data_prefix = NULL,
  measure_aux = NULL,
  targets = NULL,
  low_base = 100,
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{data}{Data}

\item{measure_column}{Character}

\item{month_col}{Character}

\item{three_month_rolling}{Boolean}

\item{last_valid_month_range_selection}{List of dates}

\item{data_prefix}{Character}

\item{measure_aux}{Data}

\item{targets}{Data}

\item{low_base}{Int}

\item{title}{Character}

\item{subtitle}{Character}
}
\value{
Highcharts object
}
\description{
Use for NPS and NES line charts over time (monthly and 3 month rolling
average); measure should be either "NPS" or "NES"
}
\examples{
\dontrun{

}
}
