% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{month_range_indicator}
\alias{month_range_indicator}
\title{Create string showing month range}
\usage{
month_range_indicator(month_range)
}
\arguments{
\item{month_range}{List of dates}
}
\value{
Character
}
\description{
Create string showing month range
}
\examples{
\dontrun{}
}
