% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{nps_group_chart}
\alias{nps_group_chart}
\title{Use for NPS groups barcharts (monthly and 3 month rolling average)}
\usage{
nps_group_chart(data, three_month_rolling = FALSE, low_base = 100, month_col)
}
\arguments{
\item{data}{Data}

\item{three_month_rolling}{Boolean}

\item{low_base}{Int}

\item{month_col}{Character}
}
\value{
Highcharts object
}
\description{
Use for NPS groups barcharts (monthly and 3 month rolling average)
}
\examples{
\dontrun{

}
}
