% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{per_s_error_vect}
\alias{per_s_error_vect}
\title{https://www.dummies.com/education/math/statistics/how-to-compare-two-population-proportions/
Why use 1.93? It seems this corresponds to a confidence interval of ~94.64\% Should it not be 1.96?}
\usage{
per_s_error_vect(base1, per1, base2, per2)
}
\arguments{
\item{base1}{Int}

\item{per1}{Numeric}

\item{base2}{Int}

\item{per2}{Numeric}
}
\value{
Character
}
\description{
https://www.dummies.com/education/math/statistics/how-to-compare-two-population-proportions/
Why use 1.93? It seems this corresponds to a confidence interval of ~94.64\% Should it not be 1.96?
}
\examples{
\dontrun{}
}
