% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{pull_value}
\alias{pull_value}
\title{Title}
\usage{
pull_value(.data, .property)
}
\arguments{
\item{.data}{Data}

\item{.property}{Character}
}
\value{
Vector of type corresponding to the data
}
\description{
Title
}
\examples{
\dontrun{

}
}
