% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{remove_line}
\alias{remove_line}
\title{Title}
\usage{
remove_line(lines, to_remove)
}
\arguments{
\item{lines}{Vector of characters}

\item{to_remove}{Character}
}
\value{
Vector of characters or NULL
}
\description{
Title
}
\examples{
\dontrun{

}
}
