% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{remove_pages}
\alias{remove_pages}
\title{Title}
\usage{
remove_pages(pages, page_menu_names = pages, backup_dirs = NULL, open = FALSE)
}
\arguments{
\item{pages}{Vector of characters}

\item{page_menu_names}{Vector of characters}

\item{backup_dirs}{Vector of filepaths}

\item{open}{Boolean}
}
\value{
Used for side effects
}
\description{
Title
}
\examples{
\dontrun{}
}
