% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{selectInputFromMonths}
\alias{selectInputFromMonths}
\title{Create select input with choices corresponding to months in data.}
\usage{
selectInputFromMonths(id, month_data, init_hide = FALSE)
}
\arguments{
\item{id}{Character}

\item{month_data}{Data}

\item{init_hide}{Boolean}
}
\value{
HTML
}
\description{
Create select input with choices corresponding to months in data.
}
\examples{
\dontrun{}
}
