% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{stacked_horizontal}
\alias{stacked_horizontal}
\title{Use for horizontal stacked barcharts showing percentages of some categorical
feature. These show a full range bar and a 3 month rolling bar for the last
month in the selected range.}
\usage{
stacked_horizontal(
  data,
  question_column,
  month_col,
  colours,
  last_valid_month_range_selection,
  responses,
  legend_names,
  low_base = 100
)
}
\arguments{
\item{data}{Data}

\item{question_column}{Character}

\item{month_col}{Character}

\item{colours}{Vector of characters}

\item{last_valid_month_range_selection}{List of dates}

\item{responses}{Variable}

\item{legend_names}{Variable}

\item{low_base}{Int}
}
\value{
Highcharts object
}
\description{
Use for horizontal stacked barcharts showing percentages of some categorical
feature. These show a full range bar and a 3 month rolling bar for the last
month in the selected range.
}
\examples{
\dontrun{

}
}
