% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{summarise_data}
\alias{summarise_data}
\title{Title}
\usage{
summarise_data(.data, .data_meta)
}
\arguments{
\item{.data}{Data}

\item{.data_meta}{Data}
}
\value{
Used for side effects
}
\description{
Title
}
\examples{
\dontrun{

}
}
