% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{suppress_errors}
\alias{suppress_errors}
\title{Selectively suppress errors using a given function to check for content in
error messages, or a character string to look for in the error message}
\usage{
suppress_errors(.expr, .f, ...)
}
\arguments{
\item{.expr}{Code}

\item{.f}{Function or character string}

\item{...}{Vector of characters}
}
\value{
Used for side effects
}
\description{
Selectively suppress errors using a given function to check for content in
error messages, or a character string to look for in the error message
}
\examples{
\dontrun{

}
}
