% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{suppress_warnings}
\alias{suppress_warnings}
\title{Selectively suppress warnings using a given function to check for content in
warning messages}
\usage{
suppress_warnings(.expr, .f, ...)
}
\arguments{
\item{.expr}{Code}

\item{.f}{Function}

\item{...}{Vector of characters}
}
\value{
Used for side effects
}
\description{
Selectively suppress warnings using a given function to check for content in
warning messages
}
\examples{
\dontrun{

}
}
