% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{timeseries}
\alias{timeseries}
\title{Time series chart}
\usage{
timeseries(
  data,
  is.percentage = FALSE,
  units = "",
  show_prev_fy = TRUE,
  prev_fy_start,
  prev_fy_end,
  this_fy_start,
  this_fy_end
)
}
\arguments{
\item{data}{Data}

\item{is.percentage}{Boolean}

\item{units}{Character}

\item{show_prev_fy}{Boolean}

\item{prev_fy_start}{Date}

\item{prev_fy_end}{Date}

\item{this_fy_start}{Date}

\item{this_fy_end}{Date}
}
\value{
Highcharts object
}
\description{
Time series chart
}
\examples{
\dontrun{}
}
