% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{updateA11yPrettyCheckboxGroup}
\alias{updateA11yPrettyCheckboxGroup}
\title{Accessibility compliant version of shinyWidgets::updatePrettyCheckboxGroup}
\usage{
updateA11yPrettyCheckboxGroup(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  choices = NULL,
  selected = NULL,
  inline = FALSE,
  choiceNames = NULL,
  choiceValues = NULL,
  prettyOptions = list()
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{choices}{The choices to set for the input object, updating choices will reset
parameters like \code{status}, \code{shape}, ... on the checkboxes, you can re-specify
(or change them) in argument \code{prettyOptions}.}

\item{selected}{The value to set for the input object.}

\item{inline}{If \code{TRUE}, render the choices inline (i.e. horizontally).}

\item{choiceNames}{The choices names to set for the input object.}

\item{choiceValues}{The choices values to set for the input object.}

\item{prettyOptions}{Arguments passed to \code{\link[shinyWidgets]{prettyCheckboxGroup}}
for styling checkboxes. This can be needed if you update choices.}
}
\value{
HTML
}
\description{
Rewrite of shinyWidgets::updatePrettyCheckboxGroup for accessibility fixes.
}
\examples{
\dontrun{
updateA11yPrettyCheckboxGroup(
  inputId = "help-used",
  choices = c("Helpdesk", "Online", "Some new help channel")
)
}
}
