% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{updateA11yPrettyOptions}
\alias{updateA11yPrettyOptions}
\title{Accessibility compliant version of shinyWidgets:::updatePrettyOptions}
\usage{
updateA11yPrettyOptions(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  choices = NULL,
  selected = NULL,
  inline = FALSE,
  type = NULL,
  choiceNames = NULL,
  choiceValues = NULL,
  prettyOptions = list()
)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{choices}{The choices to set for the input object, updating choices will
reset parameters like status, shape, ... on the checkboxes, you can re-specify
(or change them) in argument prettyOptions.}

\item{selected}{The value to set for the input object.}

\item{inline}{If TRUE, render the choices inline (i.e. horizontally).}

\item{type}{One of 'checkbox' or 'radio'.}

\item{choiceNames}{The choices names to set for the input object.}

\item{choiceValues}{The choices values to set for the input object.}

\item{prettyOptions}{Arguments passed to prettyCheckboxGroup for styling checkboxes.
This can be needed if you update choices.}
}
\value{
HTML
}
\description{
Accessibility compliant version of shinyWidgets:::updatePrettyOptions
}
\examples{
\dontrun{

}
}
