% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{user_group_select}
\alias{user_group_select}
\title{Create Shiny input for selecting NPS groups}
\usage{
user_group_select(id)
}
\arguments{
\item{id}{Character}
}
\value{
HTML
}
\description{
Create Shiny input for selecting NPS groups
}
\examples{
\dontrun{

}
}
