% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{with_plus}
\alias{with_plus}
\title{Title}
\usage{
with_plus(x, ...)
}
\arguments{
\item{x}{Numeric}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Character
}
\description{
Title
}
\examples{
with_plus(3.142, digits = 2)
}
