% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{a11yBsCollapse}
\alias{a11yBsCollapse}
\title{Accessibility compliant version of shinyBS::bsCollapse}
\usage{
a11yBsCollapse(id, open, ...)
}
\arguments{
\item{id}{\bold{Optional} You can use \code{input$id} in your Server logic to
determine which panels are open, and \code{\link[shinyBS]{updateCollapse}} to open/close
panels.}

\item{open}{The \code{value}, (or if none was supplied, the \code{title}) of
the panel(s) you want open on load.}

\item{...}{\code{\link[shinyBS]{bsCollapsePanel}} elements to include in the Collapse.}
}
\value{
HTML
}
\description{
Accessibility compliant version of shinyBS::bsCollapse
}
\examples{
\dontrun{}
}
