% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp_utils.R
\name{add_page}
\alias{add_page}
\title{Title}
\usage{
add_page(
  page_name,
  page_menu_name = page_name,
  visibility = "hidden",
  backup_dirs = NULL,
  overwrite = FALSE,
  open = FALSE
)
}
\arguments{
\item{page_name}{Character}

\item{page_menu_name}{Character}

\item{visibility}{"hidden" if not first page, "visible" for first page}

\item{backup_dirs}{Vector of filepaths}

\item{overwrite}{Boolean}

\item{open}{Boolean}
}
\value{
Used for side effects
}
\description{
Title
}
\examples{
\dontrun{}
}
