% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{line_chart}
\alias{line_chart}
\title{Use for NPS and NES line charts over time (monthly and 3 month rolling
average); measure should be either "NPS" or "NES"}
\usage{
line_chart(
  data,
  measure,
  measure_aux = NULL,
  targets = NULL,
  three_month_rolling = FALSE,
  low_base = 100,
  title = NULL,
  subtitle = NULL,
  month_col,
  last_valid_month_range_selection
)
}
\arguments{
\item{data}{Data}

\item{measure}{Character}

\item{measure_aux}{Data}

\item{targets}{Data}

\item{three_month_rolling}{Boolean}

\item{low_base}{Int}

\item{title}{Character}

\item{subtitle}{Character}

\item{month_col}{Character}

\item{last_valid_month_range_selection}{List of dates}
}
\value{
Highcharts object
}
\description{
Use for NPS and NES line charts over time (monthly and 3 month rolling
average); measure should be either "NPS" or "NES"
}
\examples{
\dontrun{

}
}
