% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{month_diff}
\alias{month_diff}
\title{Get number of months between 2 dates. A partial month counts as one month.}
\usage{
month_diff(d1, d2 = NULL)
}
\arguments{
\item{d1}{Date or character parseable as YMD date}

\item{d2}{Date or character parseable as YMD date}
}
\value{
Int
}
\description{
Get number of months between 2 dates. A partial month counts as one month.
}
\examples{
\dontrun{}
}
