% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{per_s_error}
\alias{per_s_error}
\title{Create large standard error arrow if sig diff found}
\usage{
per_s_error(base1, per1, base2, per2)
}
\arguments{
\item{base1}{Int}

\item{per1}{Numeric}

\item{base2}{Int}

\item{per2}{Numeric}
}
\value{
Character
}
\description{
You may want to avoid using this function, it is here for legacy use.
}
\examples{
per_s_error(101, 0.5, 99, 0.6)
}
