% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{pie}
\alias{pie}
\title{Pie chart showing percentage of unique responses. Use for non-rating style
questions with 3 or less possible responses only.}
\usage{
pie(
  data,
  question_column,
  colours = NULL,
  chart_title = NULL,
  low_base = 100,
  responses,
  keep_undisclosed = FALSE
)
}
\arguments{
\item{data}{Data}

\item{question_column}{Character}

\item{colours}{Vector of characters}

\item{chart_title}{Character}

\item{low_base}{Int}

\item{responses}{Variable}

\item{keep_undisclosed}{Boolean}
}
\value{
Highcharts object
}
\description{
Pie chart showing percentage of unique responses. Use for non-rating style
questions with 3 or less possible responses only.
}
\examples{
\dontrun{

}
}
