% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{stacked_vertical}
\alias{stacked_vertical}
\title{Use for vertical stacked barcharts showing percentages of some categorical feature.}
\usage{
stacked_vertical(
  data,
  columns,
  responses,
  colours,
  legend_title,
  low_base = 100,
  show_mean = FALSE,
  month_col
)
}
\arguments{
\item{data}{Data}

\item{columns}{Vector of characters}

\item{responses}{Variable}

\item{colours}{Vector of characters}

\item{legend_title}{Character}

\item{low_base}{Int}

\item{show_mean}{Boolean}

\item{month_col}{Character}
}
\value{
Plotly object
}
\description{
Use for vertical stacked barcharts showing percentages of some categorical feature.
}
\examples{
\dontrun{}
}
