% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{stacked_vertical}
\alias{stacked_vertical}
\title{Use for vertical stacked barcharts showing percentages of some categorical feature.}
\usage{
stacked_vertical(
  data,
  columns,
  month_col,
  radio_select = NULL,
  cols_start_with = TRUE,
  responses,
  keep_undisclosed = FALSE,
  colours = NULL,
  legend_title = element_blank(),
  low_base = 100,
  show_mean = FALSE
)
}
\arguments{
\item{data}{Data}

\item{columns}{Vector of characters}

\item{month_col}{Character}

\item{radio_select}{Character}

\item{cols_start_with}{Boolean}

\item{responses}{Variable}

\item{keep_undisclosed}{Boolean}

\item{colours}{Vector of characters}

\item{legend_title}{Character}

\item{low_base}{Int}

\item{show_mean}{Boolean}
}
\value{
Plotly object
}
\description{
Use for vertical stacked barcharts showing percentages of some categorical feature.
}
\examples{
\dontrun{

}
}
