% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{suppress_warnings}
\alias{suppress_warnings}
\title{Selectively suppress warnings using a given function to check for content in
warning messages, or a character string to look for in the warning message}
\usage{
suppress_warnings(.expr, .f, ...)
}
\arguments{
\item{.expr}{Code}

\item{.f}{Function or character string}

\item{...}{Vector of characters}
}
\value{
Used for side effects
}
\description{
The purpose of this is to prevent known, i.e. expected, warnings from
appearing. This is helpful as you then know if you see red text when running
some code that it is something unexpected!
}
\examples{
suppress_warnings({
    as.numeric(c("18", "30", "50+", "345,678"))
  },
  endsWith,
  "by coercion"
)
}
